requirements.txt
PySimpleGUI==4.60.4
clipspy==1.0.0